Sample a degree corrected overlapping stochastic blockmodel graph

Sample a degree corrected overlapping stochastic blockmodel graph

dcOverlapping(theta, pi, B, ...)

Arguments

theta
TODO

pi
TODO

B
TODO

...
Arguments passed on to fastRG

X

An n by k_1 matrix.

S

A k_1 by k_2 matrix.

Y

A d by k_2 matrix. Defaults to X.

avgDeg

When specified, the expected average degree of nodes in the output poisson multi-graph. When poissonEdges = FALSE, the resulting graph will have lower average degree due to lack of multiple-edges. When the graph is sparse, the expected number of edges for the Poisson multi-graph and Bernoulli graph are nearly the same. Defaults to NULL, such that no scaling occurs.

simple

When TRUE indicates that you want to work with undirected graphs where self-loops and multi-edges are prohibited. Accomplishes this by setting directed = FALSE, selfLoops = FALSE, and PoissonEdges = FALSE, and then ignoring arguments directed, selfLoops and PoissonEdges. Defaults to FALSE.

PoissonEdges

Logical indicating whether or not multi-edges are allowed. Defaults to TRUE, which keeps multi-edges and produces a multi-graph. When FALSE, only single edges are allowed, resulting in a graph. See details for some additional comments. Effected by simple argument.

directed

Logical indicating whether or not the graph should be directed. Default is directed = TRUE. When directed = FALSE, symmetrizes S internally. When X = Y (which is the default when no Y is specified), this results in a symmetric adjacency matrix as output. When avgDeg is specified and the desired graph is directed, the average degree scaling is on the out-degree of each node (or the row sums if you prefer to think in terms of the adjacency matrix). Effected by the simple argument.

selfLoops

Logical indicating whether edges are allowed from a node back to itself. Defaults to TRUE. When FALSE, sampling proceeds normally, and then self-loops are removed after sampling is completed. Effected by the simple argument.

returnEdgeList

Logical indicating whether to return an edgelist rather than an adjacency matrix.

Value

A random Poisson (or Bernoulli) dot product graph. By default, returns a Matrix::sparseMatrix() in CSC format (i.e. of abstract class CsparseMatrix). When the graph is undirected, the sparseMatrix will also be symmetric.

If returnEdgeList = TRUE, instead returns the graph as an edgelist in matrix format, where the first column of the matrix contains two columns from and to . The from column contains the index of the outgoing edge in this case, and the to column contains the index of the incoming edge.

Details

TODO: write the model out in detail here

to remove DC, set theta=rep(1,n) $$ xi = \theta_i * z_i, where [z_i]_j ~ bernoulli(pi_j) $$

$$ \lambda_{ij} = xi' B xj $$

probability of \(i\) connecting to \(j\): \(1 - exp(-\lambda_{ij})\)

See also

Examples


# TODO