Create an directed erdos renyi object

Create an directed erdos renyi object

directed_erdos_renyi(n, ..., p = NULL)

Arguments

n
Number of nodes in graph.

...
Arguments passed on to directed_factor_model

expected_in_degree

If specified, the desired expected in degree of the graph. Specifying expected_in_degree simply rescales S to achieve this. Defaults to NULL. Specify only one of expected_in_degree, expected_out_degree, and expected_density.

expected_out_degree

If specified, the desired expected out degree of the graph. Specifying expected_out_degree simply rescales S to achieve this. Defaults to NULL. Specify only one of expected_in_degree, expected_out_degree, and expected_density.

p
Probability of an edge between any two nodes. You must specify either p, expected_in_degree, or expected_out_degree.

Value

Never returns Poisson edges.

See also

Examples

set.seed(87)

er <- directed_erdos_renyi(n = 10, p = 0.1)
er

#> Directed Factor Model
#> ---------------------
#> 
#> Incoming Nodes (n): 10
#> Incoming Rank (k1): 1
#> Outgoing Rank (k2): 1
#> Outgoing Nodes (d): 10
#> 
#> X: 10 x 1 [dgeMatrix] 
#> S: 1 x 1 [ddiMatrix] 
#> Y: 10 x 1 [dgeMatrix] 
#> 
#> Expected edges: 11
#> Expected density: 0.10536
#> Expected in degree: 1.1
#> Expected out degree: 1.1

big_er <- directed_erdos_renyi(n = 10^6, expected_in_degree = 5)
big_er

#> Directed Factor Model
#> ---------------------
#> 
#> Incoming Nodes (n): 1000000
#> Incoming Rank (k1): 1
#> Outgoing Rank (k2): 1
#> Outgoing Nodes (d): 1000000
#> 
#> X: 1000000 x 1 [dgeMatrix] 
#> S: 1 x 1 [ddiMatrix] 
#> Y: 1000000 x 1 [dgeMatrix] 
#> 
#> Expected edges: 5e+06
#> Expected density: 0
#> Expected in degree: 5
#> Expected out degree: 5

A <- sample_sparse(er)
A

#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                          
#>  [1,] . . . . . . . . . .
#>  [2,] . . . . . . . 1 . .
#>  [3,] . . . . . 1 . . . .
#>  [4,] . . . . 1 . 1 . 1 .
#>  [5,] . . . . . . . . . .
#>  [6,] . . . . . . 1 . . .
#>  [7,] . . . 1 . . . . . .
#>  [8,] . . . . . . . . . 1
#>  [9,] . . . . . . 1 . . .
#> [10,] 1 . . 1 . . . . . .