Sample a random dot product graph as an igraph graph

There are two steps to using the fastRG package. First, you must parameterize a random dot product graph by specifying its expected adjacency matrix. Use functions such as dcsbm(), sbm(), etc, to perform this specification. Then, use sample_igraph() to generate a random graph, represented as an igraph::igraph(), with that expectation.

sample_igraph(factor_model, ..., poisson_edges = TRUE, allow_self_loops = TRUE)

# S3 method for undirected_factor_model
sample_igraph(factor_model, ..., poisson_edges = TRUE, allow_self_loops = TRUE)

# S3 method for directed_factor_model
sample_igraph(factor_model, ..., poisson_edges = TRUE, allow_self_loops = TRUE)

# S3 method for undirected_erdos_renyi
sample_igraph(
  factor_model,
  ...,
  poisson_edges = FALSE,
  allow_self_loops = TRUE
)

# S3 method for undirected_erdos_renyi
sample_igraph(
  factor_model,
  ...,
  poisson_edges = FALSE,
  allow_self_loops = TRUE
)

Arguments

factor_model
A directed_factor_model() or undirected_factor_model().

...
Ignored. Do not use.

poisson_edges
Logical indicating whether or not multiple edges are allowed to form between a pair of nodes. Defaults to TRUE. When FALSE, sampling proceeds as usual, and duplicate edges are removed afterwards. See Section 2.3 of Rohe et al (2017) for additional details.

allow_self_loops
Logical indicating whether or not nodes should be allowed to form edges with themselves. Defaults to TRUE. When FALSE, sampling proceeds as usual, and self-loops are removed afterwards.

Value

An igraph::igraph() graph object.

Details

This function implements the fastRG algorithm as described in Rohe et al (2017). Please see the paper (which is short and open access!!) for details.

References

Rohe, Karl, Jun Tao, Xintian Han, and Norbert Binkiewicz. 2017. "A Note on Quickly Sampling a Sparse Matrix with Low Rank Expectation." Journal of Machine Learning Research; 19(77):1-13, 2018. http://www.jmlr.org/papers/v19/17-128.html

See also

Examples

set.seed(27)

##### undirected examples ----------------------------

n <- 1000
k <- 5

X <- matrix(rpois(n = n * k, 1), nrow = n)
S <- matrix(runif(n = k * k, 0, .1), nrow = k)

ufm <- undirected_factor_model(
  X, S,
  expected_density = 0.1
)

ufm

#> Undirected Factor Model
#> -----------------------
#> 
#> Nodes (n): 1000
#> Rank (k): 5
#> 
#> X: 1000 x 5 [dgeMatrix] 
#> S: 5 x 5 [dgeMatrix] 
#> 
#> Expected edges: 50000
#> Expected degree: 50
#> Expected density: 0.05

### sampling graphs as edgelists ----------------------

edgelist <- sample_edgelist(ufm)
edgelist

#> # A tibble: 24,951 x 2
#>     from    to
#>    <int> <int>
#>  1   691   831
#>  2   309   988
#>  3   469   371
#>  4   109   954
#>  5    74   397
#>  6   400   982
#>  7    47   952
#>  8   116   600
#>  9   384   652
#> 10   854   818
#> # … with 24,941 more rows

### sampling graphs as sparse matrices ----------------

A <- sample_sparse(ufm)

inherits(A, "dsCMatrix")  # TRUE

#> [1] TRUE

class(A[2, 1])            # "numeric" -- i.e. A has double data type

#> [1] "numeric"

isSymmetric(A)            # TRUE

#> [1] TRUE

B <- sample_sparse(ufm, poisson_edges = FALSE)

inherits(B, "dsCMatrix")  # TRUE

#> [1] TRUE

isSymmetric(B)            # TRUE

#> [1] TRUE

### sampling graphs as igraph graphs ------------------

sample_igraph(ufm)

#> IGRAPH 589d431 U--- 1000 25027 -- 
#> + edges from 589d431:
#>  [1] 758--982 702--912 113--509 326--428 158--185 638--707 110--497  86--538
#>  [9] 130--362  75--486 388--845 184--270 231--979 442--624 756--965 115--278
#> [17] 313--773 597--780  22--508 552--597 531--927 929--931 331--639 120--215
#> [25] 435--966   1--780 554--722 580--695 233--284 105--708  65--213 597--720
#> [33]  13--708 445--750 172--526 171--702   1--326 668--929   1--770 325--527
#> [41] 159--956 233--773 758--929 163--562 339--753 761--929 755--929  45--766
#> [49] 690--769 268--354 115--855 184--209  55--381 947--952 449--460 282--831
#> [57] 821--844  27--872 702--839 318--783 435--872 133--402 547--796 486--756
#> [65] 496--554 161--453  77--982 471--600 595--638  73--968 341--447 262--731
#> + ... omitted several edges

### sampling graphs as tidygraph graphs ---------------

sample_tidygraph(ufm)

#> # A tbl_graph: 986 nodes and 24921 edges
#> #
#> # An undirected multigraph with 1 component
#> #
#> # Node Data: 986 x 1 (active)
#>   name 
#>   <chr>
#> 1 485  
#> 2 355  
#> 3 368  
#> 4 722  
#> 5 914  
#> 6 375  
#> # … with 980 more rows
#> #
#> # Edge Data: 24,921 x 2
#>    from    to
#>   <int> <int>
#> 1     1    66
#> 2     2   182
#> 3     3   429
#> # … with 24,918 more rows

sample_tidygraph(ufm, poisson_edges = FALSE)

#> # A tbl_graph: 986 nodes and 24570 edges
#> #
#> # An undirected multigraph with 1 component
#> #
#> # Node Data: 986 x 1 (active)
#>   name 
#>   <chr>
#> 1 983  
#> 2 658  
#> 3 253  
#> 4 844  
#> 5 221  
#> 6 720  
#> # … with 980 more rows
#> #
#> # Edge Data: 24,570 x 2
#>    from    to
#>   <int> <int>
#> 1     1   266
#> 2     2   350
#> 3     3   135
#> # … with 24,567 more rows

##### directed examples ----------------------------

n2 <- 10000

k1 <- 5
k2 <- 3

d <- 5000

X <- matrix(rpois(n = n2 * k1, 1), nrow = n2)
S <- matrix(runif(n = k1 * k2, 0, .1), nrow = k1, ncol = k2)
Y <- matrix(rexp(n = k2 * d, 1), nrow = d)

fm <- directed_factor_model(X, S, Y, expected_in_degree = 50)
fm

#> Directed Factor Model
#> ---------------------
#> 
#> Incoming Nodes (n): 10000
#> Incoming Rank (k1): 5
#> Outgoing Rank (k2): 3
#> Outgoing Nodes (d): 5000
#> 
#> X: 10000 x 5 [dgeMatrix] 
#> S: 5 x 3 [dgeMatrix] 
#> Y: 5000 x 3 [dgeMatrix] 
#> 
#> Expected edges: 5e+05
#> Expected density: 0.01
#> Expected in degree: 50
#> Expected out degree: 100

### sampling graphs as edgelists ----------------------

edgelist2 <- sample_edgelist(fm)
edgelist2

#> # A tibble: 500,592 x 2
#>     from    to
#>    <int> <int>
#>  1   351  4865
#>  2  6482  2676
#>  3  2742  2316
#>  4  9406  4523
#>  5  1412  1706
#>  6  1819  2960
#>  7  4000  3484
#>  8  3986  3948
#>  9  8587  2823
#> 10  5481  4134
#> # … with 500,582 more rows

### sampling graphs as sparse matrices ----------------

A2 <- sample_sparse(fm)

inherits(A2, "dgCMatrix") # TRUE

#> [1] TRUE

isSymmetric(A2)           # FALSE

#> [1] FALSE

B2 <- sample_sparse(fm, poisson_edges = FALSE)

inherits(B2, "dgCMatrix")  # TRUE

#> [1] TRUE

isSymmetric(B2)            # TRUE

#> [1] FALSE

### sampling graphs as igraph graphs ------------------

sample_igraph(fm)

#> IGRAPH c9f7247 D--- 10000 500118 -- 
#> + edges from c9f7247:
#>  [1] 2082->3989 1055->3991  804->1998 8479->4066 1458->2982 2811->4904
#>  [7] 2929->1865 4684-> 925 7940->  24 5506->4450 8351->1442 2496->1323
#> [13] 3991-> 564 9403->3582 6450-> 536 5448->4120 3445->1770 4089->2303
#> [19] 9381->1625  199-> 677 7827->3224  727->1989 7823->2606 8035->3149
#> [25] 2496->3813 1507->2378 3248-> 838 2688->3513 5485->3895 7531->3719
#> [31] 6582->3325 6422->3656 9118->2276 4069->4621 7281->1454 7116->2901
#> [37] 9030->4520 9195->1620 4458-> 364 6071->2233 7873->4780 9952->2510
#> [43] 2106->4167  305->3456  280->3010 2578->4165 1033-> 391 7411->1403
#> [49] 3460-> 793 7787->2557 2282->3786 3207->4748   15->3408 1113->2056
#> + ... omitted several edges

### sampling graphs as tidygraph graphs ---------------

sample_tidygraph(fm)

#> # A tbl_graph: 9968 nodes and 499069 edges
#> #
#> # A directed multigraph with 1 component
#> #
#> # Node Data: 9,968 x 1 (active)
#>   name 
#>   <chr>
#> 1 3632 
#> 2 2604 
#> 3 4309 
#> 4 3249 
#> 5 1382 
#> 6 5229 
#> # … with 9,962 more rows
#> #
#> # Edge Data: 499,069 x 2
#>    from    to
#>   <int> <int>
#> 1     1  4225
#> 2     2  6136
#> 3     3  9327
#> # … with 499,066 more rows

sample_tidygraph(fm, poisson_edges = FALSE)

#> # A tbl_graph: 9968 nodes and 496828 edges
#> #
#> # A directed multigraph with 1 component
#> #
#> # Node Data: 9,968 x 1 (active)
#>   name 
#>   <chr>
#> 1 5956 
#> 2 332  
#> 3 3660 
#> 4 8314 
#> 5 7285 
#> 6 9882 
#> # … with 9,962 more rows
#> #
#> # Edge Data: 496,828 x 2
#>    from    to
#>   <int> <int>
#> 1     1  6565
#> 2     2  1028
#> 3     3  3649
#> # … with 496,825 more rows