Create an undirected factor model graph

An undirected factor model graph is an undirected generalized Poisson random dot product graph. The edges in this graph are assumed to be independent and Poisson distributed. The graph is parameterized by its expected adjacency matrix, which is E[A|X] = X S X'. We do not recommend that casual users use this function, see instead dcsbm() and related functions, which will formulate common variants of the stochastic blockmodels as undirected factor models with lots of helpful input validation.

Usage

undirected_factor_model(
  X,
  S,
  ...,
  expected_degree = NULL,
  expected_density = NULL
)

Arguments

X

A matrix() or Matrix() representing real-valued latent node positions. Entries must be positive.

S

A matrix() or Matrix() mixing matrix. S is symmetrized if it is not already, as this is the undirected case. Entries must be positive.

...

Ignored. Must be empty.

expected_degree

If specified, the desired expected degree of the graph. Specifying expected_degree simply rescales S to achieve this. Defaults to NULL. Do not specify both expected_degree and expected_density at the same time.

expected_density

If specified, the desired expected density of the graph. Specifying expected_density simply rescales S to achieve this. Defaults to NULL. Do not specify both expected_degree and expected_density at the same time.

Value

An undirected_factor_model S3 class based on a list with the following elements:

X: The latent positions as a Matrix() object.

S: The mixing matrix as a Matrix() object.

n: The number of nodes in the network.

k: The rank of expectation matrix. Equivalently, the dimension of the latent node position vectors.

Examples

n <- 10000
k <- 5

X <- matrix(rpois(n = n * k, 1), nrow = n)
S <- matrix(runif(n = k * k, 0, .1), nrow = k)

ufm <- undirected_factor_model(X, S)
ufm
#> Undirected Factor Model
#> -----------------------
#> 
#> Nodes (n): 10000
#> Rank (k): 5
#> 
#> X: 10000 x 5 [dgeMatrix] 
#> S: 5 x 5 [dgeMatrix] 
#> 
#> Expected edges: 226530209
#> Expected degree: 22653
#> Expected density: 2.2653

sane_ufm <- undirected_factor_model(X, S, expected_degree = 50)
sane_ufm
#> Undirected Factor Model
#> -----------------------
#> 
#> Nodes (n): 10000
#> Rank (k): 5
#> 
#> X: 10000 x 5 [dgeMatrix] 
#> S: 5 x 5 [dgeMatrix] 
#> 
#> Expected edges: 5e+05
#> Expected degree: 50
#> Expected density: 0.005