Create an undirected Chung-Lu object

To specify a Chung-Lu graph, you must specify the degree-heterogeneity parameters (via n or theta). We provide reasonable defaults to enable rapid exploration or you can invest the effort for more control over the model parameters. We strongly recommend setting the expected_degree or expected_density argument to avoid large memory allocations associated with sampling large, dense graphs.

Usage

chung_lu(n = NULL, theta = NULL, ..., sort_nodes = TRUE)

Arguments

n

(degree heterogeneity) The number of nodes in the graph. Use when you don't want to specify the degree-heterogeneity parameters theta by hand. When n is specified, theta is randomly generated from a LogNormal(2, 1) distribution. This is subject to change, and may not be reproducible. n defaults to NULL. You must specify either n or theta, but not both.

theta

(degree heterogeneity) A numeric vector explicitly specifying the degree heterogeneity parameters. This implicitly determines the number of nodes in the resulting graph, i.e. it will have length(theta) nodes. Must be positive. Setting to a vector of ones recovers an erdos renyi graph. Defaults to NULL. You must specify either n or theta, but not both.

...

Arguments passed on to undirected_factor_model

expected_degree

If specified, the desired expected degree of the graph. Specifying expected_degree simply rescales S to achieve this. Defaults to NULL. Do not specify both expected_degree and expected_density at the same time.

expected_density

If specified, the desired expected density of the graph. Specifying expected_density simply rescales S to achieve this. Defaults to NULL. Do not specify both expected_degree and expected_density at the same time.

sort_nodes

Logical indicating whether or not to sort the nodes so that they are grouped by block and by theta. Useful for plotting. Defaults to TRUE.

Value

An undirected_chung_lu S3 object, a subclass of dcsbm().

See also

Examples

set.seed(27)

cl <- chung_lu(n = 1000, expected_density = 0.01)
#> Generating random degree heterogeneity parameters `theta` from a LogNormal(2, 1) distribution. This distribution may change in the future. Explicitly set `theta` for reproducible results.
cl
#> Undirected Degree-Corrected Stochastic Blockmodel
#> -------------------------------------------------
#> 
#> Nodes (n): 1000 (arranged by block)
#> Blocks (k): 1
#> 
#> Traditional DCSBM parameterization:
#> 
#> Block memberships (z): 1000 [factor] 
#> Degree heterogeneity (theta): 1000 [numeric] 
#> Block probabilities (pi): 1 [numeric] 
#> 
#> Factor model parameterization:
#> 
#> X: 1000 x 1 [dgeMatrix] 
#> S: 1 x 1 [ddiMatrix] 
#> 
#> Expected edges: 10000
#> Expected degree: 10
#> Expected density: 0.01

theta <- round(stats::rlnorm(100, 2))

cl2 <- chung_lu(
  theta = theta,
  expected_degree = 5
)

cl2
#> Undirected Degree-Corrected Stochastic Blockmodel
#> -------------------------------------------------
#> 
#> Nodes (n): 100 (arranged by block)
#> Blocks (k): 1
#> 
#> Traditional DCSBM parameterization:
#> 
#> Block memberships (z): 100 [factor] 
#> Degree heterogeneity (theta): 100 [numeric] 
#> Block probabilities (pi): 1 [numeric] 
#> 
#> Factor model parameterization:
#> 
#> X: 100 x 1 [dgeMatrix] 
#> S: 1 x 1 [ddiMatrix] 
#> 
#> Expected edges: 500
#> Expected degree: 5
#> Expected density: 0.05

edgelist <- sample_edgelist(cl)
edgelist
#> # A tibble: 5,077 × 2
#>     from    to
#>    <int> <int>
#>  1    17    49
#>  2    34    80
#>  3     5     6
#>  4   648   930
#>  5    88   666
#>  6   122   277
#>  7     6    18
#>  8    65   708
#>  9    48   608
#> 10    59   309
#> # … with 5,067 more rows