Sample a random dot product graph as an igraph graph

There are two steps to using the fastRG package. First, you must parameterize a random dot product graph by sampling the latent factors. Use functions such as dcsbm(), sbm(), etc, to perform this specification. Then, use sample_*() functions to generate a random graph in your preferred format.

Usage

# S3 method for directed_erdos_renyi
sample_igraph(factor_model, ..., allow_self_loops = TRUE)

sample_igraph(factor_model, ..., poisson_edges = TRUE, allow_self_loops = TRUE)

# S3 method for undirected_factor_model
sample_igraph(factor_model, ..., poisson_edges = TRUE, allow_self_loops = TRUE)

# S3 method for directed_factor_model
sample_igraph(factor_model, ..., poisson_edges = TRUE, allow_self_loops = TRUE)

# S3 method for undirected_erdos_renyi
sample_igraph(
  factor_model,
  ...,
  poisson_edges = FALSE,
  allow_self_loops = TRUE
)

# S3 method for undirected_erdos_renyi
sample_igraph(
  factor_model,
  ...,
  poisson_edges = FALSE,
  allow_self_loops = TRUE
)

Arguments

factor_model

A directed_factor_model() or undirected_factor_model().

...

Ignored. Do not use.

allow_self_loops

Logical indicating whether or not nodes should be allowed to form edges with themselves. Defaults to TRUE. When FALSE, sampling proceeds allowing self-loops, and these are then removed after the fact.

poisson_edges

Logical indicating whether or not multiple edges are allowed to form between a pair of nodes. Defaults to TRUE. When FALSE, sampling proceeds as usual, and duplicate edges are removed afterwards. See Section 2.3 of Rohe et al (2017) for additional details.

Value

An igraph::igraph() object that is possibly a multigraph (that is, we take there to be multiple edges rather than weighted edges).

When factor_model is undirected:

- the graph is undirected and one-mode.

When factor_model is directed and square:

- the graph is directed and one-mode.

When factor_model is directed and rectangular:

- the graph is undirected and bipartite.

Note that working with bipartite graphs in igraph is more complex than working with one-mode graphs.

Details

This function implements the fastRG algorithm as described in Rohe et al (2017). Please see the paper (which is short and open access!!) for details.

References

Rohe, Karl, Jun Tao, Xintian Han, and Norbert Binkiewicz. 2017. "A Note on Quickly Sampling a Sparse Matrix with Low Rank Expectation." Journal of Machine Learning Research; 19(77):1-13, 2018. https://www.jmlr.org/papers/v19/17-128.html

See also

Examples

library(igraph)
library(tidygraph)

set.seed(27)

##### undirected examples ----------------------------

n <- 100
k <- 5

X <- matrix(rpois(n = n * k, 1), nrow = n)
S <- matrix(runif(n = k * k, 0, .1), nrow = k)

# S will be symmetrized internal here, or left unchanged if
# it is already symmetric

ufm <- undirected_factor_model(
  X, S,
  expected_density = 0.1
)

ufm
#> Undirected Factor Model
#> -----------------------
#> 
#> Nodes (n): 100
#> Rank (k): 5
#> 
#> X: 100 x 5 [dgeMatrix] 
#> S: 5 x 5 [dgeMatrix] 
#> 
#> Expected edges: 1000
#> Expected degree: 10
#> Expected density: 0.1

### sampling graphs as edgelists ----------------------

edgelist <- sample_edgelist(ufm)
edgelist
#> # A tibble: 505 × 2
#>     from    to
#>    <int> <int>
#>  1    66    71
#>  2    83    87
#>  3    37    87
#>  4    92    95
#>  5    14    94
#>  6    54    85
#>  7    11    76
#>  8    45    57
#>  9    57    74
#> 10    11    22
#> # … with 495 more rows

### sampling graphs as sparse matrices ----------------

A <- sample_sparse(ufm)

inherits(A, "dsCMatrix")
#> [1] TRUE
isSymmetric(A)
#> [1] TRUE
dim(A)
#> [1] 100 100

B <- sample_sparse(ufm, poisson_edges = FALSE)

inherits(B, "dsCMatrix")
#> [1] TRUE
isSymmetric(B)
#> [1] TRUE
dim(B)
#> [1] 100 100

### sampling graphs as igraph graphs ------------------

sample_igraph(ufm)
#> IGRAPH 5506f6d UN-- 100 526 -- 
#> + attr: name (v/c)
#> + edges from 5506f6d (vertex names):
#>  [1] 12--25  14--45  22--50  6 --68  1 --16  17--92  15--77  6 --44  94--94 
#> [10] 22--82  24--56  61--76  39--95  79--82  25--77  22--75  30--32  27--32 
#> [19] 22--85  30--44  54--87  25--47  32--75  55--100 83--84  28--80  64--71 
#> [28] 94--55  71--80  57--88  61--87  57--5   8 --40  74--96  75--76  75--23 
#> [37] 25--56  30--80  75--44  34--52  75--40  24--88  24--33  1 --38  64--26 
#> [46] 29--98  23--66  25--98  61--97  57--58  27--56  61--100 16--62  15--69 
#> [55] 74--80  94--65  6 --64  15--26  71--84  17--59  22--83  44--85  39--84 
#> [64] 74--23  94--73  27--13  38--76  16--92  22--99  57--81  16--43  14--48 
#> + ... omitted several edges

### sampling graphs as tidygraph graphs ---------------

sample_tidygraph(ufm)
#> # A tbl_graph: 100 nodes and 438 edges
#> #
#> # An undirected multigraph with 2 components
#> #
#> # Node Data: 100 × 1 (active)
#>    name
#>   <int>
#> 1     1
#> 2     2
#> 3     3
#> 4     4
#> 5     5
#> 6     6
#> # … with 94 more rows
#> #
#> # Edge Data: 438 × 2
#>    from    to
#>   <int> <int>
#> 1     1    76
#> 2     1    94
#> 3    92    92
#> # … with 435 more rows

sample_tidygraph(ufm, poisson_edges = FALSE)
#> # A tbl_graph: 100 nodes and 441 edges
#> #
#> # An undirected multigraph with 3 components
#> #
#> # Node Data: 100 × 1 (active)
#>    name
#>   <int>
#> 1     1
#> 2     2
#> 3     3
#> 4     4
#> 5     5
#> 6     6
#> # … with 94 more rows
#> #
#> # Edge Data: 441 × 2
#>    from    to
#>   <int> <int>
#> 1    80    87
#> 2    45    80
#> 3    38    85
#> # … with 438 more rows

##### directed examples ----------------------------

n2 <- 100

k1 <- 5
k2 <- 3

d <- 50

X <- matrix(rpois(n = n2 * k1, 1), nrow = n2)
S <- matrix(runif(n = k1 * k2, 0, .1), nrow = k1, ncol = k2)
Y <- matrix(rexp(n = k2 * d, 1), nrow = d)

fm <- directed_factor_model(X, S, Y, expected_in_degree = 2)
fm
#> Directed Factor Model
#> ---------------------
#> 
#> Incoming Nodes (n): 100
#> Incoming Rank (k1): 5
#> Outgoing Rank (k2): 3
#> Outgoing Nodes (d): 50
#> 
#> X: 100 x 5 [dgeMatrix] 
#> S: 5 x 3 [dgeMatrix] 
#> Y: 50 x 3 [dgeMatrix] 
#> 
#> Expected edges: 100
#> Expected density: 0.02
#> Expected in degree: 2
#> Expected out degree: 1

### sampling graphs as edgelists ----------------------

edgelist2 <- sample_edgelist(fm)
edgelist2
#> # A tibble: 93 × 2
#>     from    to
#>    <int> <int>
#>  1     4    23
#>  2    37    20
#>  3    40    27
#>  4    31    18
#>  5    52    47
#>  6    64    11
#>  7    74    23
#>  8    95    46
#>  9    85    39
#> 10    79    30
#> # … with 83 more rows

### sampling graphs as sparse matrices ----------------

A2 <- sample_sparse(fm)

inherits(A2, "dgCMatrix")
#> [1] TRUE
isSymmetric(A2)
#> [1] FALSE
dim(A2)
#> [1] 100  50

B2 <- sample_sparse(fm, poisson_edges = FALSE)

inherits(B2, "dgCMatrix")
#> [1] TRUE
isSymmetric(B2)
#> [1] FALSE
dim(B2)
#> [1] 100  50

### sampling graphs as igraph graphs ------------------

# since the number of rows and the number of columns
# in `fm` differ, we will get a bipartite igraph here

# creating the bipartite igraph is slow relative to other
# sampling -- if this is a blocker for
# you please open an issue and we can investigate speedups

dig <- sample_igraph(fm)
is_bipartite(dig)
#> [1] TRUE

### sampling graphs as tidygraph graphs ---------------

sample_tidygraph(fm, poisson_edges = FALSE)
#> # A tbl_graph: 150 nodes and 97 edges
#> #
#> # A bipartite simple graph with 59 components
#> #
#> # Node Data: 150 × 1 (active)
#>   type 
#>   <lgl>
#> 1 FALSE
#> 2 FALSE
#> 3 FALSE
#> 4 FALSE
#> 5 FALSE
#> 6 FALSE
#> # … with 144 more rows
#> #
#> # Edge Data: 97 × 2
#>    from    to
#>   <int> <int>
#> 1     6   101
#> 2    43   101
#> 3    44   101
#> # … with 94 more rows