Sample a random dot product graph as an igraph graph

There are two steps to using the fastRG package. First, you must parameterize a random dot product graph by sampling the latent factors. Use functions such as dcsbm(), sbm(), etc, to perform this specification. Then, use sample_*() functions to generate a random graph in your preferred format.

# S3 method for directed_erdos_renyi
sample_igraph(factor_model, ..., allow_self_loops = TRUE)

sample_igraph(factor_model, ..., poisson_edges = TRUE, allow_self_loops = TRUE)

# S3 method for undirected_factor_model
sample_igraph(factor_model, ..., poisson_edges = TRUE, allow_self_loops = TRUE)

# S3 method for directed_factor_model
sample_igraph(factor_model, ..., poisson_edges = TRUE, allow_self_loops = TRUE)

# S3 method for undirected_erdos_renyi
sample_igraph(
  factor_model,
  ...,
  poisson_edges = FALSE,
  allow_self_loops = TRUE
)

# S3 method for undirected_erdos_renyi
sample_igraph(
  factor_model,
  ...,
  poisson_edges = FALSE,
  allow_self_loops = TRUE
)

Arguments

factor_model
A directed_factor_model() or undirected_factor_model().

...
Ignored. Do not use.

allow_self_loops
Logical indicating whether or not nodes should be allowed to form edges with themselves. Defaults to TRUE. When FALSE, sampling proceeds allowing self-loops, and these are then removed after the fact.

poisson_edges
Logical indicating whether or not multiple edges are allowed to form between a pair of nodes. Defaults to TRUE. When FALSE, sampling proceeds as usual, and duplicate edges are removed afterwards. See Section 2.3 of Rohe et al (2017) for additional details.

Value

An igraph::igraph() object that is possibly a multigraph (that is, we take there to be multiple edges rather than weighted edges).

When factor_model is undirected:

- the graph is undirected and one-mode.

When factor_model is directed and square:

- the graph is directed and one-mode.

When factor_model is directed and rectangular:

- the graph is undirected and bipartite.

Note that working with bipartite graphs in igraph is more complex than working with one-mode graphs.

Details

This function implements the fastRG algorithm as described in Rohe et al (2017). Please see the paper (which is short and open access!!) for details.

References

Rohe, Karl, Jun Tao, Xintian Han, and Norbert Binkiewicz. 2017. "A Note on Quickly Sampling a Sparse Matrix with Low Rank Expectation." Journal of Machine Learning Research; 19(77):1-13, 2018. http://www.jmlr.org/papers/v19/17-128.html

See also

Examples

library(igraph)
library(tidygraph)

set.seed(27)

##### undirected examples ----------------------------

n <- 1000
k <- 5

X <- matrix(rpois(n = n * k, 1), nrow = n)
S <- matrix(runif(n = k * k, 0, .1), nrow = k)

# S will be symmetrized internal here, or left unchanged if
# it is already symmetric

ufm <- undirected_factor_model(
  X, S,
  expected_density = 0.1
)

ufm

#> Undirected Factor Model
#> -----------------------
#> 
#> Nodes (n): 1000
#> Rank (k): 5
#> 
#> X: 1000 x 5 [dgeMatrix] 
#> S: 5 x 5 [dgeMatrix] 
#> 
#> Expected edges: 1e+05
#> Expected degree: 100
#> Expected density: 0.1

### sampling graphs as edgelists ----------------------

edgelist <- sample_edgelist(ufm)
edgelist

#> # A tibble: 49,931 x 2
#>     from    to
#>    <int> <int>
#>  1   298   691
#>  2   309   629
#>  3   143   469
#>  4   109   953
#>  5    74   478
#>  6    61   400
#>  7    47    70
#>  8   116   453
#>  9   161   384
#> 10   229   854
#> # … with 49,921 more rows

### sampling graphs as sparse matrices ----------------

A <- sample_sparse(ufm)

inherits(A, "dsCMatrix")

#> [1] TRUE

isSymmetric(A)

#> [1] TRUE

dim(A)

#> [1] 1000 1000

B <- sample_sparse(ufm, poisson_edges = FALSE)

inherits(B, "dsCMatrix")

#> [1] TRUE

isSymmetric(B)

#> [1] TRUE

dim(B)

#> [1] 1000 1000

### sampling graphs as igraph graphs ------------------

sample_igraph(ufm)

#> IGRAPH ab0d3fd UN-- 986 49931 -- 
#> + attr: name (v/c)
#> + edges from ab0d3fd (vertex names):
#>  [1] 248--784 284--466 262--362 435--714 122--134 143--388 375--760 977--982
#>  [9] 85 --808 384--722 496--903 238--268 80 --147 858--880 86 --207 864--952
#> [17] 854--929 691--784 623--903 127--240 399--612 566--687 447--823 36 --112
#> [25] 736--857 331--589 195--766 399--777 502--637 404--474 102--844 184--799
#> [33] 47 --235 864--354 17 --219 16 --298 735--947 147--839 538--766 714--868
#> [41] 112--857 399--17  197--753 287--562 474--515 73 --883 911--914 784--922
#> [49] 100--961 111--634 375--135 743--998 399--594 204--833 206--745 225--826
#> [57] 278--654 188--697 14 --915 27 --75  135--465 354--182 362--818 426--695
#> + ... omitted several edges

### sampling graphs as tidygraph graphs ---------------

sample_tidygraph(ufm)

#> # A tbl_graph: 986 nodes and 49790 edges
#> #
#> # An undirected multigraph with 1 component
#> #
#> # Node Data: 986 x 1 (active)
#>   name 
#>   <chr>
#> 1 346  
#> 2 399  
#> 3 497  
#> 4 552  
#> 5 463  
#> 6 441  
#> # … with 980 more rows
#> #
#> # Edge Data: 49,790 x 2
#>    from    to
#>   <int> <int>
#> 1     1   492
#> 2     2   281
#> 3     3   403
#> # … with 49,787 more rows

sample_tidygraph(ufm, poisson_edges = FALSE)

#> # A tbl_graph: 986 nodes and 46035 edges
#> #
#> # An undirected multigraph with 1 component
#> #
#> # Node Data: 986 x 1 (active)
#>   name 
#>   <chr>
#> 1 186  
#> 2 157  
#> 3 235  
#> 4 163  
#> 5 257  
#> 6 209  
#> # … with 980 more rows
#> #
#> # Edge Data: 46,035 x 2
#>    from    to
#>   <int> <int>
#> 1     1   448
#> 2     2   540
#> 3     3   971
#> # … with 46,032 more rows

##### directed examples ----------------------------

n2 <- 10000

k1 <- 5
k2 <- 3

d <- 5000

X <- matrix(rpois(n = n2 * k1, 1), nrow = n2)
S <- matrix(runif(n = k1 * k2, 0, .1), nrow = k1, ncol = k2)
Y <- matrix(rexp(n = k2 * d, 1), nrow = d)

fm <- directed_factor_model(X, S, Y, expected_in_degree = 50)
fm

#> Directed Factor Model
#> ---------------------
#> 
#> Incoming Nodes (n): 10000
#> Incoming Rank (k1): 5
#> Outgoing Rank (k2): 3
#> Outgoing Nodes (d): 5000
#> 
#> X: 10000 x 5 [dgeMatrix] 
#> S: 5 x 3 [dgeMatrix] 
#> Y: 5000 x 3 [dgeMatrix] 
#> 
#> Expected edges: 250000
#> Expected density: 0.005
#> Expected in degree: 50
#> Expected out degree: 25

### sampling graphs as edgelists ----------------------

edgelist2 <- sample_edgelist(fm)
edgelist2

#> # A tibble: 249,807 x 2
#>     from    to
#>    <int> <int>
#>  1  3607  4964
#>  2  4580  4094
#>  3  1876  3836
#>  4   994  4531
#>  5  7665  3080
#>  6  9038   496
#>  7   663  2173
#>  8  6465   673
#>  9  7074  3877
#> 10  2922  2767
#> # … with 249,797 more rows

### sampling graphs as sparse matrices ----------------

A2 <- sample_sparse(fm)

inherits(A2, "dgCMatrix")

#> [1] TRUE

isSymmetric(A2)

#> [1] FALSE

dim(A2)

#> [1] 10000  5000

B2 <- sample_sparse(fm, poisson_edges = FALSE)

inherits(B2, "dgCMatrix")

#> [1] TRUE

isSymmetric(B2)

#> [1] FALSE

dim(B2)

#> [1] 10000  5000

### sampling graphs as igraph graphs ------------------

# since the number of rows and the number of columns
# in `fm` differ, we will get a bipartite igraph here

# creating the bipartite igraph is slow relative to other
# sampling -- if this is a blocker for
# you please open an issue and we can investigate speedups

dig <- sample_igraph(fm)

is_bipartite(dig)

#> [1] TRUE

### sampling graphs as tidygraph graphs ---------------

sample_tidygraph(fm)

#> # A tbl_graph: 15000 nodes and 250234 edges
#> #
#> # A bipartite multigraph with 89 components
#> #
#> # Node Data: 15,000 x 1 (active)
#>   type 
#>   <lgl>
#> 1 FALSE
#> 2 FALSE
#> 3 FALSE
#> 4 FALSE
#> 5 FALSE
#> 6 FALSE
#> # … with 14,994 more rows
#> #
#> # Edge Data: 250,234 x 2
#>    from    to
#>   <int> <int>
#> 1   307 10001
#> 2   355 10001
#> 3   736 10001
#> # … with 250,231 more rows

sample_tidygraph(fm, poisson_edges = FALSE)

#> # A tbl_graph: 15000 nodes and 248484 edges
#> #
#> # A bipartite simple graph with 85 components
#> #
#> # Node Data: 15,000 x 1 (active)
#>   type 
#>   <lgl>
#> 1 FALSE
#> 2 FALSE
#> 3 FALSE
#> 4 FALSE
#> 5 FALSE
#> 6 FALSE
#> # … with 14,994 more rows
#> #
#> # Edge Data: 248,484 x 2
#>    from    to
#>   <int> <int>
#> 1   585 10001
#> 2   979 10001
#> 3  1065 10001
#> # … with 248,481 more rows