Sample a random dot product graph as a tidygraph graph

There are two steps to using the fastRG package. First, you must parameterize a random dot product graph by specifying its expected adjacency matrix. Use functions such as dcsbm(), sbm(), etc, to perform this specification. Then, use sample_tidygraph() to generate a random graph, represented as an tidygraph::tbl_graph() with that expectation.

sample_tidygraph(
  factor_model,
  ...,
  poisson_edges = TRUE,
  allow_self_loops = TRUE
)

# S3 method for undirected_factor_model
sample_tidygraph(
  factor_model,
  ...,
  poisson_edges = TRUE,
  allow_self_loops = TRUE
)

# S3 method for directed_factor_model
sample_tidygraph(
  factor_model,
  ...,
  poisson_edges = TRUE,
  allow_self_loops = TRUE
)

# S3 method for undirected_erdos_renyi
sample_tidygraph(
  factor_model,
  ...,
  poisson_edges = FALSE,
  allow_self_loops = TRUE
)

# S3 method for undirected_erdos_renyi
sample_tidygraph(
  factor_model,
  ...,
  poisson_edges = FALSE,
  allow_self_loops = TRUE
)

Arguments

factor_model
A directed_factor_model(), undirected_factor_model(), or one of the many subclasses.

...
Ignored. Do not use.

poisson_edges
Logical indicating whether or not multiple edges are allowed to form between a pair of nodes. Defaults to TRUE. When FALSE, sampling proceeds as usual, and duplicate edges are removed afterwards. See Section 2.3 of Rohe et al (2017) for additional details.

allow_self_loops
Logical indicating whether or not nodes should be allowed to form edges with themselves. Defaults to TRUE. When FALSE, sampling proceeds as usual, and self-loops are removed afterwards.

Value

A tidygraph::tbl_graph() object.

Details

This function implements the fastRG algorithm as described in Rohe et al (2017). Please see the paper (which is short and open access!!) for details.

References

Rohe, Karl, Jun Tao, Xintian Han, and Norbert Binkiewicz. 2017. "A Note on Quickly Sampling a Sparse Matrix with Low Rank Expectation." Journal of Machine Learning Research; 19(77):1-13, 2018. http://www.jmlr.org/papers/v19/17-128.html

See also

Examples

set.seed(27)

##### undirected examples ----------------------------

n <- 1000
k <- 5

X <- matrix(rpois(n = n * k, 1), nrow = n)
S <- matrix(runif(n = k * k, 0, .1), nrow = k)

ufm <- undirected_factor_model(
  X, S,
  expected_density = 0.1
)

ufm

#> Undirected Factor Model
#> -----------------------
#> 
#> Nodes (n): 1000
#> Rank (k): 5
#> 
#> X: 1000 x 5 [dgeMatrix] 
#> S: 5 x 5 [dgeMatrix] 
#> 
#> Expected edges: 50000
#> Expected degree: 50
#> Expected density: 0.05

### sampling graphs as edgelists ----------------------

edgelist <- sample_edgelist(ufm)
edgelist

#> # A tibble: 49,931 x 2
#>     from    to
#>    <int> <int>
#>  1   691   298
#>  2   309   629
#>  3   469   143
#>  4   109   953
#>  5    74   478
#>  6   400    61
#>  7    47    70
#>  8   116   453
#>  9   384   161
#> 10   854   229
#> # … with 49,921 more rows

### sampling graphs as sparse matrices ----------------

A <- sample_sparse(ufm)

inherits(A, "dsCMatrix")  # TRUE

#> [1] TRUE

class(A[2, 1])            # "numeric" -- i.e. A has double data type

#> [1] "numeric"

isSymmetric(A)            # TRUE

#> [1] TRUE

B <- sample_sparse(ufm, poisson_edges = FALSE)

inherits(B, "dsCMatrix")  # TRUE

#> [1] TRUE

isSymmetric(B)            # TRUE

#> [1] TRUE

### sampling graphs as igraph graphs ------------------

sample_igraph(ufm)

#> IGRAPH 503ba13 U--- 1000 49931 -- 
#> + edges from 503ba13:
#>  [1] 248--784 284--466 262--362 435--714 122--134 143--388 375--760 977--982
#>  [9]  85--808 384--722 496--903 238--268  80--147 858--880  86--207 864--952
#> [17] 854--929 691--784 623--903 127--240 399--612 566--687 447--823  36--112
#> [25] 736--857 331--589 195--766 399--777 502--637 404--474 102--844 184--799
#> [33]  47--235 354--864  17--219  16--298 735--947 147--839 538--766 714--868
#> [41] 112--857  17--399 197--753 287--562 474--515  73--883 911--914 784--922
#> [49] 100--961 111--634 135--375 743--998 399--594 204--833 206--745 225--826
#> [57] 278--654 188--697  14--915  27-- 75 135--465 182--354 362--818 426--695
#> [65] 554--658 144--203 634--855 257--320  47--203 655--821  95--890 117--807
#> + ... omitted several edges

### sampling graphs as tidygraph graphs ---------------

sample_tidygraph(ufm)

#> # A tbl_graph: 986 nodes and 49790 edges
#> #
#> # An undirected multigraph with 1 component
#> #
#> # Node Data: 986 x 1 (active)
#>   name 
#>   <chr>
#> 1 888  
#> 2 399  
#> 3 497  
#> 4 552  
#> 5 887  
#> 6 789  
#> # … with 980 more rows
#> #
#> # Edge Data: 49,790 x 2
#>    from    to
#>   <int> <int>
#> 1     1    72
#> 2     2   627
#> 3     3   307
#> # … with 49,787 more rows

sample_tidygraph(ufm, poisson_edges = FALSE)

#> # A tbl_graph: 986 nodes and 47702 edges
#> #
#> # An undirected multigraph with 1 component
#> #
#> # Node Data: 986 x 1 (active)
#>   name 
#>   <chr>
#> 1 186  
#> 2 157  
#> 3 996  
#> 4 788  
#> 5 753  
#> 6 209  
#> # … with 980 more rows
#> #
#> # Edge Data: 47,702 x 2
#>    from    to
#>   <int> <int>
#> 1     1   105
#> 2     2   525
#> 3     3   593
#> # … with 47,699 more rows

##### directed examples ----------------------------

n2 <- 10000

k1 <- 5
k2 <- 3

d <- 5000

X <- matrix(rpois(n = n2 * k1, 1), nrow = n2)
S <- matrix(runif(n = k1 * k2, 0, .1), nrow = k1, ncol = k2)
Y <- matrix(rexp(n = k2 * d, 1), nrow = d)

fm <- directed_factor_model(X, S, Y, expected_in_degree = 50)
fm

#> Directed Factor Model
#> ---------------------
#> 
#> Incoming Nodes (n): 10000
#> Incoming Rank (k1): 5
#> Outgoing Rank (k2): 3
#> Outgoing Nodes (d): 5000
#> 
#> X: 10000 x 5 [dgeMatrix] 
#> S: 5 x 3 [dgeMatrix] 
#> Y: 5000 x 3 [dgeMatrix] 
#> 
#> Expected edges: 5e+05
#> Expected density: 0.01
#> Expected in degree: 50
#> Expected out degree: 100

### sampling graphs as edgelists ----------------------

edgelist2 <- sample_edgelist(fm)
edgelist2

#> # A tibble: 499,727 x 2
#>     from    to
#>    <int> <int>
#>  1  3607  4512
#>  2  4580  4583
#>  3  1876  1989
#>  4   994  2945
#>  5  7665  2347
#>  6  9038  4582
#>  7   663  1491
#>  8  6465   709
#>  9  7074  3241
#> 10  2922   101
#> # … with 499,717 more rows

### sampling graphs as sparse matrices ----------------

A2 <- sample_sparse(fm)

inherits(A2, "dgCMatrix") # TRUE

#> [1] TRUE

isSymmetric(A2)           # FALSE

#> [1] FALSE

B2 <- sample_sparse(fm, poisson_edges = FALSE)

inherits(B2, "dgCMatrix")  # TRUE

#> [1] TRUE

isSymmetric(B2)            # TRUE

#> [1] FALSE

### sampling graphs as igraph graphs ------------------

sample_igraph(fm)

#> IGRAPH 28ffe27 D--- 10000 499235 -- 
#> + edges from 28ffe27:
#>  [1]   51->2177 2924-> 227 7482->4484 8282->3871 9902-> 297 5457-> 392
#>  [7] 9669->4585 7374->2869 5666->3145 6725->1926 8604->1820 9027-> 254
#> [13] 9292-> 297 6115-> 295  478->3650 8177->1148 7798-> 209 7423->3280
#> [19] 7996->3386 8079->4774 8164->3990 9028->1349   71->2178 3493->3837
#> [25] 3857->4148 8878->4802 8124->4582 2493->1795 5355->4073  858->4563
#> [31] 3213->3147 4126->4666 1928->1350  304->4582    5->1543 2272->1543
#> [37] 2786-> 128  243-> 126 5580->1246 7315->2671 7021->2396 7174->3938
#> [43] 6855->1628 9799-> 415 8442->4865  321->3856 6313-> 447 4539->2828
#> [49] 2726->4013 3173->1597  959->3589 5150->1651 5603->4268  775->1600
#> + ... omitted several edges

### sampling graphs as tidygraph graphs ---------------

sample_tidygraph(fm)

#> # A tbl_graph: 9966 nodes and 501048 edges
#> #
#> # A directed multigraph with 1 component
#> #
#> # Node Data: 9,966 x 1 (active)
#>   name 
#>   <chr>
#> 1 2095 
#> 2 1390 
#> 3 1512 
#> 4 8988 
#> 5 5340 
#> 6 3962 
#> # … with 9,960 more rows
#> #
#> # Edge Data: 501,048 x 2
#>    from    to
#>   <int> <int>
#> 1     1  1148
#> 2     2  1153
#> 3     3  8969
#> # … with 501,045 more rows

sample_tidygraph(fm, poisson_edges = FALSE)

#> # A tbl_graph: 9966 nodes and 495220 edges
#> #
#> # A directed multigraph with 1 component
#> #
#> # Node Data: 9,966 x 1 (active)
#>   name 
#>   <chr>
#> 1 8587 
#> 2 976  
#> 3 9527 
#> 4 3685 
#> 5 7136 
#> 6 3246 
#> # … with 9,960 more rows
#> #
#> # Edge Data: 495,220 x 2
#>    from    to
#>   <int> <int>
#> 1     1  8102
#> 2     2  1224
#> 3     3  1072
#> # … with 495,217 more rows